fts_project: A means to search "projects" and tasks.
One limitation of the free text search is that it works on individual tables,
and doesn't appear to support searching on the join of two tables.

In the case of Projects; the project name and description is part of its
analytic account.  Project-specific fields are part of the project itself, so
we need to query across account_analytic_account and project_project to be
thorough.  I don't know how to achieve this with the present FTS framework.